Main config set up
