+++ index.html
@@ -16,9 +16,8 @@
     <link rel="stylesheet" href="./style.min.css" />
   </head>
 
-  <body>
+  <body class="bg-grey-lightest">
     <div id="app"></div>
-    <div class="basic">Testing</div>
 
     <script type="text/javascript" src="./main.min.js"></script>
   </body>
