+++ js/components/Footer.vue
@@ -1,9 +1,19 @@
 <template>
-  <div class="footer">Attribution etc<div>
+  <div class="card mt-6">
+    <p>
+      This is an
+      <a href="http://xkcd.com/936/">XKCD inspired</a> password generator.
+    </p>
+    <p>*Saving options will utilise LocalStorage. De-selecting will remove the options from LocalStorage.</p>
+    <p>
+      Maintained by
+      <a href="https://github.com/FSMaxB">FSMaxB</a> on
+      <a href="https://github.com/FSMaxB/correcthorsebatterystaple">GitHub</a>
+    </p>
+    <p>
+      Forked from
+      <a href="http://twitter.com/geekyjohn">JVDL</a> (on
+      <a href="https://bitbucket.org/jvdl/correcthorsebatterystaple">Bitbucket</a>)
+    </p>
+  </div>
 </template>
-
-<style lang="scss">
-.footer {
-  background: blue;
-}
-</style>
