+++ js/components/Main.vue
@@ -1,3 +1,6 @@
 <template>
-  <div class="basic">Card is in here</div>
+  <div class="card">
+    <h1>Correct Horse Battery Staple</h1>
+    <h4>Secure password generator to help keep you safer online</h4>
+  </div>
 </template>
