+++ js/main.js
@@ -4,5 +4,6 @@ import App from "./App.vue"
 
 new Vue({
   el: "#app",
+  template: `<App />`,
   components: { App },
 })
