+++ tailwind.js
@@ -14,7 +14,7 @@ let colors = {
   grey: "#b8c2cc",
   "grey-light": "#dae1e7",
   "grey-lighter": "#f1f5f8",
-  "grey-lightest": "#f8fafc",
+  "grey-lightest": "#F9FBFC",
   white: "#ffffff",
 
   "red-darkest": "#3b0d0c",
@@ -585,6 +585,7 @@ module.exports = {
     lg: "0 15px 30px 0 rgba(0,0,0,0.11), 0 5px 15px 0 rgba(0,0,0,0.08)",
     inner: "inset 0 2px 4px 0 rgba(0,0,0,0.06)",
     outline: "0 0 0 3px rgba(52,144,220,0.5)",
+    card: "16px 16px 64px rgba(0,0,0,0.08)",
     none: "none",
   },
